Typo: remove duplicated --dry-run in --help
The current verbose help has this list of flags:

```
      flags:
...
        --dry-run,-n             don’t do anything, just print
...
        --dry-run,-n             don’t execute, just print
```

Which has the `--dry-run` flag duplicated. This commit removes the second instance. Since the wording was slightly different for each of the instances, I kept the wording and the ordering as is used for the "less verbose" version of `--help`, which looks like:

```
flags:
  --sync,-S       sync and update environment packages
  --env,-E        inject local environment
  --dry-run,-n    don’t do anything, just print
```