Fix colorization in shellcode errors
This because deno’s console %c colorizer always removes color codes if stdout is not a tty. This doesn't apply to us in these cases because we are outputting to stderr and that *is* a tty.

This is what happens when you are using a product that is mainly targeted at web-devs and not cli-devs.